[UPDT] Recuring templates 'base_accounting_kit'
