+++ base_accounting_kit/__manifest__.py
@@ -22,7 +22,7 @@
 
 {
     'name': 'Odoo 13 Full Accounting Kit',
-    'version': '13.0.2.4.2',
+    'version': '13.0.3.4.2',
     'category': 'Accounting',
     'summary': """ Asset and Budget Management,
                  Accounting Reports, PDC, Lock dates, 
@@ -53,6 +53,7 @@
         'data/account_pdc_data.xml',
         'data/followup_levels.xml',
         'data/account_asset_data.xml',
+        'data/recurring_entry_cron.xml',
         'views/reports_config_view.xml',
         'views/accounting_menu.xml',
         'views/credit_limit_view.xml',
@@ -80,7 +81,6 @@
         'wizard/account_bank_book_wizard_view.xml',
         'wizard/account_cash_book_wizard_view.xml',
         'wizard/account_day_book_wizard_view.xml',
-        'wizard/recurring_payments_wizard.xml',
         'report/report_financial.xml',
         'report/general_ledger_report.xml',
         'report/report_journal_audit.xml',
