+++ base_accounting_kit/models/recurring_payments.py
@@ -23,7 +23,8 @@ from datetime import datetime, date
 
 from dateutil.relativedelta import relativedelta
 
-from odoo import models, fields, api
+from odoo import models, fields, api, _
+from odoo.exceptions import UserError
 
 
 class FilterRecurringEntries(models.Model):
@@ -89,8 +90,92 @@ class RecurringPayments(models.Model):
                                 store=True, required=True)
     company_id = fields.Many2one('res.company',
                                  default=lambda l: l.env.user.company_id.id)
+    recurring_lines = fields.One2many('account.recurring.entries.line', 'tmpl_id')
 
     @api.onchange('partner_id')
     def onchange_partner_id(self):
         if self.partner_id.property_account_receivable_id:
             self.credit_account = self.partner_id.property_account_payable_id
+
+    @api.model
+    def _cron_generate_entries(self):
+        data = self.env['account.recurring.payments'].search(
+            [('state', '=', 'running')])
+        entries = self.env['account.move'].search(
+            [('recurring_ref', '!=', False)])
+        journal_dates = []
+        journal_codes = []
+        remaining_dates = []
+        for entry in entries:
+            journal_dates.append(str(entry.date))
+            if entry.recurring_ref:
+                journal_codes.append(str(entry.recurring_ref))
+        today = datetime.today()
+        for line in data:
+            if line.date:
+                recurr_dates = []
+                start_date = datetime.strptime(str(line.date), '%Y-%m-%d')
+                while start_date <= today:
+                    recurr_dates.append(str(start_date.date()))
+                    if line.recurring_period == 'days':
+                        start_date += relativedelta(
+                            days=line.recurring_interval)
+                    elif line.recurring_period == 'weeks':
+                        start_date += relativedelta(
+                            weeks=line.recurring_interval)
+                    elif line.recurring_period == 'months':
+                        start_date += relativedelta(
+                            months=line.recurring_interval)
+                    else:
+                        start_date += relativedelta(
+                            years=line.recurring_interval)
+                for rec in recurr_dates:
+                    recurr_code = str(line.id) + '/' + str(rec)
+                    if recurr_code not in journal_codes:
+                        remaining_dates.append({
+                            'date': rec,
+                            'template_name': line.name,
+                            'amount': line.amount,
+                            'tmpl_id': line.id,
+                        })
+        child_ids = self.recurring_lines.create(remaining_dates)
+        if not child_ids:
+            raise UserError(_("There is no remaining payments"))
+        for line in child_ids:
+            tmpl_id = line.tmpl_id
+            recurr_code = str(tmpl_id.id) + '/' + str(line.date)
+            line_ids = [(0, 0, {
+                'account_id': tmpl_id.credit_account.id,
+                'partner_id': tmpl_id.partner_id.id,
+                'credit': line.amount,
+                'analytic_account_id': tmpl_id.analytic_account_id.id,
+            }), (0, 0, {
+                'account_id': tmpl_id.debit_account.id,
+                'partner_id': tmpl_id.partner_id.id,
+                'debit': line.amount,
+                'analytic_account_id': tmpl_id.analytic_account_id.id,
+            })]
+            vals = {
+                'date': line.date,
+                'recurring_ref': recurr_code,
+                'company_id': self.env.user.company_id.id,
+                'journal_id': tmpl_id.journal_id.id,
+                'ref': line.template_name,
+                'narration': 'Recurring entry',
+                'line_ids': line_ids
+            }
+            move_id = self.env['account.move'].create(vals)
+            if tmpl_id.journal_state == 'posted':
+                move_id.post()
+
+
+    class GetAllRecurringEntries(models.TransientModel):
+        _name = 'account.recurring.entries.line'
+        _description = 'Account Recurring Entries Line'
+
+        date = fields.Date('Date')
+        template_name = fields.Char('Name')
+        amount = fields.Float('Amount')
+        tmpl_id = fields.Many2one('account.recurring.payments', string='id')
+
+
