+++ base_accounting_kit/views/account_asset_views.xml
@@ -161,7 +161,7 @@
                             <field name="value" widget="monetary" options="{'currency_field': 'currency_id'}" help="Gross value of asset"/>
                             <field name="salvage_value" widget="monetary" options="{'currency_field': 'currency_id'}" attrs="{'invisible': [('type','=','sale')]}"/>
                             <field name="value_residual" widget="monetary" options="{'currency_field': 'currency_id'}"/>
-                            <field name="partner_id" string="Vendor" domain="[('supplier', '=', True)]"/>
+                            <field name="partner_id" string="Vendor" domain="[('supplier_rank', '>', 0)]"/>
                             <field name="invoice_id" string="Invoice" attrs="{'readonly': [('state', '!=', 'draft')]}" options="{'no_create': True}"/>
                         </group>
                     </group>
