+++ base_accounting_kit/views/recurring_payments_view.xml
@@ -4,10 +4,10 @@
 
         <!--Recurring Templates Form view-->
         <record id="account_recurring_payments_form_view" model="ir.ui.view">
-            <field name="name">Recurring Payments</field>
+            <field name="name">Recurring Template</field>
             <field name="model">account.recurring.payments</field>
             <field name="arch" type="xml">
-                <form string="Recurring Payments">
+                <form string="Recurring Template">
                     <header>
                         <field name="state" widget="statusbar" clickable="1"/>
                     </header>
@@ -63,10 +63,10 @@
 
         <!--Recurring Templates Tree View-->
         <record id="account_recurring_payments_tree_view" model="ir.ui.view">
-            <field name="name">Recurring Payments</field>
+            <field name="name">Recurring Templates</field>
             <field name="model">account.recurring.payments</field>
             <field name="arch" type="xml">
-                <tree string="Recurring Payments">
+                <tree string="Recurring Templates">
                     <field name="name"/>
                     <field name="partner_id"/>
                     <field name="debit_account"/>
@@ -82,7 +82,7 @@
 
         <!--Action for Recurring Templates-->
         <record id="action_account_recurring_payments_view" model="ir.actions.act_window">
-            <field name="name">Recurring Payments</field>
+            <field name="name">Recurring Templates</field>
             <field name="res_model">account.recurring.payments</field>
             <field name="type">ir.actions.act_window</field>
             <field name="view_mode">tree,form</field>
