--- base_accounting_kit/wizard/recurring_payments_wizard.xml
@@ -1,62 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<odoo>
-    <data>
-
-        <!--Recurring Payments Wizard View-->
-        <record id="account_recurring_payments_wizard_view" model="ir.ui.view">
-            <field name="name">Generate Payments</field>
-            <field name="model">recurring.payments.wizard</field>
-            <field name="arch" type="xml">
-                <form string="Generate Payments">
-                    <div class="text-muted">
-                        Only recurring templates which are in the running stage will be considered.
-                    </div>
-                    <group>
-                        <group>
-                            <field name="date_from"/>
-                        </group>
-                        <group>
-                            <field name="recurring_tmpl_ids" widget="many2many_tags"/>
-                        </group>
-                    </group>
-
-                    <notebook>
-                        <page name="remaining_entries" string="Remaining Entries">
-                            <field name="recurring_lines">
-                                <tree string="Remaining Payments" create="false" editable="top">
-                                    <field name="date"/>
-                                    <field name="template_name"/>
-                                    <field name="amount"/>
-                                    <field name="tmpl_id" invisible="1"/>
-                                </tree>
-                            </field>
-                        </page>
-                    </notebook>
-                    <footer>
-                        <button name="generate_payment" string="Generate Entries" type="object" class="oe_highlight"/>
-                        <button string="Cancel" special="cancel"/>
-                    </footer>
-                </form>
-            </field>
-        </record>
-
-        <!--Action for Recurring Payments Wizard View-->
-        <record id="action_account_recurring_payments_wizard_view" model="ir.actions.act_window">
-            <field name="name">Recurring Payments</field>
-            <field name="res_model">recurring.payments.wizard</field>
-            <field name="type">ir.actions.act_window</field>
-            <field name="view_mode">form</field>
-            <field name="view_id" ref="account_recurring_payments_wizard_view"/>
-            <field name="target">new</field>
-        </record>
-
-        <menuitem id="account_recurring_payments_child2" name="Recurring Entries"
-                  action="action_account_recurring_payments_wizard_view"
-                  parent="account.menu_finance_receivables" sequence="16"
-                  groups="account.group_account_user,account.group_account_manager"/>
-        <menuitem id="account_recurring_payments_child3" name="Recurring Entries"
-                  action="action_account_recurring_payments_wizard_view"
-                  parent="account.menu_finance_payables" sequence="20"
-                  groups="account.group_account_user,account.group_account_manager"/>
-    </data>
-</odoo>
